Refactor: extract Card from RepoCard, merge RepoCard into Search Component
