+++ App.js
@@ -11,7 +11,6 @@ import type { Element } from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { enableScreens } from 'react-native-screens';
-import Icon from 'react-native-vector-icons/Octicons';
 
 import useStoredList from './hooks/useStoredList';
 import { RepoContext } from './context';
@@ -19,7 +18,6 @@ import Home from './Components/Home';
 import Error from './Components/Error';
 import { LIMIT } from './constants';
 
-Icon.loadFont();
 enableScreens(true);
 
 const Stack = createNativeStackNavigator();
